+++ pagnn/pagnn.py
@@ -1,6 +1,69 @@
 import torch
 
 
+def get_linear_layers(net):
+    linear_layers = []
+    for child in net.children():
+        if type(child) == torch.nn.Linear:
+            linear_layers.append(child)
+    return linear_layers
+
+
+def count_neurons(net, return_layers=False):
+    first = None
+    extra = 0
+    last = 0
+    layers = 0
+    linear_layer_list = get_linear_layers(net)
+    if len(linear_layer_list) <= 0:
+        linear_layer_list = [net]
+    for layer in linear_layer_list:
+        layers += 1
+        in_neurons = layer.in_features
+        if first is None:
+            first = in_neurons
+        else:
+            extra += in_neurons 
+        last = layer.out_features
+
+    if return_layers:
+        return first, extra, last, layers
+    return first, extra, last
+
+
+def import_ffnn(ffnn, activation):
+    """Ingest a FFNN into the PAGNN architecture"""
+
+    # create equivalent adjacency structure
+    first, extra, last, layers = count_neurons(ffnn, return_layers=True)
+    pagnn = PAGNNLayer(first, last, extra, steps=layers, activation=activation, retain_state=False)
+
+    # import synaptic weightings
+    pagnn.zero_params()
+    last_i = 0
+    seen_output_neurons = 0
+    for i, layer in enumerate(get_linear_layers(ffnn)):
+        in_neurons = layer.in_features
+        out_neurons = layer.out_features
+
+        pW = pagnn.weight
+        pb = pagnn.bias
+        lW = layer.weight
+        lb = layer.bias
+        
+        new_last_i = last_i + in_neurons
+
+        pW.data[last_i:last_i+in_neurons, new_last_i:new_last_i+out_neurons] = lW.T
+        if lb is not None:
+            pb.data[new_last_i:new_last_i+out_neurons] = lb
+
+        last_i = new_last_i
+
+        seen_output_neurons += out_neurons
+
+    return pagnn
+
+
 def _pagnn_op(state, weight, bias=None):
     if state.dim() == 2 and bias is not None:
         # fused op is marginally faster
@@ -18,7 +81,11 @@ class PAGNNLayer(torch.nn.Module):
     def __init__(self, input_neurons, output_neurons, extra_neurons, steps=1, sparsity=0, retain_state=True, activation=None):
         super(PAGNNLayer, self).__init__()
 
-        assert input_neurons > 0 and output_neurons > 0 and extra_neurons >= 0 and sparsity >= 0 and steps >= 1
+        assert input_neurons > 0 
+        assert output_neurons > 0 
+        assert extra_neurons >= 0 
+        assert sparsity >= 0 
+        assert steps >= 1
 
         if sparsity > 0:
             raise NotImplemented()
@@ -106,6 +173,34 @@ class PAGNNLayer(torch.nn.Module):
             return self.state[-self._output_neurons:]
         return self.state[:, -self._output_neurons:]
 
+
+    torch.no_grad()
+    def _load_from_state_dict(self, state_dict, prefix, local_metadata, strict,
+                              missing_keys, unexpected_keys, error_msgs):
+        use_super_load = False
+        tensors = []
+
+        for name, param in self.named_parameters():
+            key = prefix + name
+            if key in state_dict:
+                input_param = state_dict[key]
+                
+                # if shapes are equal, can defer to default super loader
+                if input_param.shape == param.shape:
+                    use_super_load = True
+                    break
+
+                tensors.append(input_param)
+
+        if use_super_load:
+            super()._load_from_state_dict(state_dict, prefix, local_metadata, strict,
+                                          missing_keys, unexpected_keys, error_msgs)
+        else:
+            W, b = tensors
+            net = torch.nn.Linear(W.shape[1], W.shape[0])
+            new_pagnn = import_ffnn(net, None)
+            self.weight.data = new_pagnn.weight.data
+            self.bias.data = new_pagnn.bias.data
 
     def extra_repr(self):
         return 'input_neurons=%i, output_neurons=%i, extra_neurons=%i' % (self._input_neurons, self._output_neurons, self._extra_neurons)
